+++ Link.Discord.Bots.Tests/Link.Discord.Bots.Tests.csproj
@@ -5,6 +5,7 @@
     <Nullable>enable</Nullable>
     <LangVersion>latest</LangVersion>
     <IsPackable>false</IsPackable>
+    <RootNamespace>Link.Discord.Bots</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
